Modify policy document resource format
The resource has been updated to allow access to all resources of
the specific API that's being called (but only if the token verifies).

This change makes it possible to cache the policy document (i.e. "
authorizer.resultTtlInSeconds") properly and without any undesired
behavior.

If the entire method ARN is used as a resource, a user will get a 403
when calling the API with different HTTP verbs (e.g HTTP GET or POST)
in the time the policy document has been cached.
For example, when the user makes a GET request with a valid token, and
caching is enabled, the policy document will be cached for X seconds.
If in that time the same user makes a different request (like POST), with
a valid token, the policy would have been cached for the resource "*/GET",
disallowing the other request for the time the policy is cached.